requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
werkzeug>=2.0.0
pillow>=8.0.0
